requirements.txt
discord.py  
wavelink    
lavalink
